requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.3.0
